Linted; commented
