+++ packages/contracts/test/contracts/ovm/execution-manager/ExecutionManager.recover-eoa-address.spec.ts
@@ -15,7 +15,7 @@ import {
   getWallets,
   makeAddressResolver,
   deployAndRegister,
-  AddressResolverMapping
+  AddressResolverMapping,
 } from '../../../test-helpers'
 
 /* Logging */
@@ -45,11 +45,7 @@ describe('Execution Manager -- Recover EOA Address', () => {
       'ExecutionManager',
       {
         factory: ExecutionManager,
-        params: [
-          resolver.addressResolver.address,
-          NULL_ADDRESS,
-          GAS_LIMIT
-        ]
+        params: [resolver.addressResolver.address, NULL_ADDRESS, GAS_LIMIT],
       }
     )
   })
