+++ packages/contracts/test/contracts/ovm/execution-manager/ExecutionManager.storage-opcodes.spec.ts
@@ -13,7 +13,7 @@ import {
   encodeRawArguments,
   makeAddressResolver,
   deployAndRegister,
-  AddressResolverMapping
+  AddressResolverMapping,
 } from '../../../test-helpers'
 
 /* Logging */
@@ -48,11 +48,7 @@ describe('ExecutionManager -- Storage opcodes', () => {
       'ExecutionManager',
       {
         factory: ExecutionManager,
-        params: [
-          resolver.addressResolver.address,
-          NULL_ADDRESS,
-          GAS_LIMIT
-        ]
+        params: [resolver.addressResolver.address, NULL_ADDRESS, GAS_LIMIT],
       }
     )
   })
