+++ packages/contracts/test/test-helpers/index.ts
@@ -6,4 +6,4 @@ export * from './ethereum-helpers'
 export * from './ovm-helpers'
 export * from './trie-helpers'
 export * from './wallet-helpers'
-export * from './resolution-helpers'
+export * from './resolution'
