+++ packages/contracts/test/test-helpers/resolution/index.ts
@@ -1,0 +1,2 @@
+export * from './resolver'
+export { AddressResolverMapping } from './types'
